move wasm impl for SystemInstruction to system-instruction crate
